+++ srcs/ft_engine.c
@@ -6,13 +6,13 @@
 /*   By: adda-sil <adda-sil@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/17 19:16:36 by adda-sil          #+#    #+#             */
-/*   Updated: 2019/11/27 22:17:31 by adda-sil         ###   ########.fr       */
+/*   Updated: 2019/11/28 22:51:07 by adda-sil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-int	ft_draw_renderer(t_game *game, t_ipos pos, unsigned long int color)
+int	ft_draw_renderer(t_game *game, t_ipos pos, unsigned int color)
 {
 	// if (pos.x < game->map.mini.width && pos.y < game->map.mini.height)
 	// 	return (0);
